refactor: split utils.ts into validation/url + utils/errors + utils/html (Plan E)
