+++ src/auth/account-manager.ts
@@ -9,7 +9,7 @@ import { getLogger } from "@logtape/logtape";
 import { z } from "zod";
 import { MAX_RESPONSE_SIZE, REQUEST_TIMEOUT } from "../config.js";
 import { readJsonWithLimit } from "../utils/fetch-helpers.js";
-import { validateExternalUrl } from "../utils.js";
+import { validateExternalUrl } from "../validation/url.js";
 
 const logger = getLogger("activitypub-mcp:account-manager");
