+++ src/discovery/instance-discovery.ts
@@ -4,8 +4,8 @@ import { fileURLToPath } from "node:url";
 import { getLogger } from "@logtape/logtape";
 import { MAX_RESPONSE_SIZE, REQUEST_TIMEOUT, USER_AGENT } from "../config.js";
 import { readJsonWithLimit } from "../utils/fetch-helpers.js";
-import { validateExternalUrl } from "../utils.js";
 import { DomainSchema } from "../validation/schemas.js";
+import { validateExternalUrl } from "../validation/url.js";
 
 const logger = getLogger("activitypub-mcp");
