+++ src/discovery/webfinger.ts
@@ -2,8 +2,8 @@ import { getLogger } from "@logtape/logtape";
 import { z } from "zod";
 import { CACHE_MAX_SIZE, CACHE_TTL, REQUEST_TIMEOUT, USER_AGENT } from "../config.js";
 import { LRUCache } from "../utils/lru-cache.js";
-import { validateExternalUrl } from "../utils.js";
 import { ActorIdentifierSchema } from "../validation/schemas.js";
+import { validateExternalUrl } from "../validation/url.js";
 
 const logger = getLogger("activitypub-mcp");
