+++ src/mcp/tools-export.ts
@@ -13,7 +13,8 @@ import { remoteClient } from "../activitypub/remote-client.js";
 import type { RateLimiter } from "../resilience/rate-limiter.js";
 import { validateActorIdentifier, validateDomain } from "../server/index.js";
 import { performanceMonitor } from "../telemetry/performance-monitor.js";
-import { formatErrorWithSuggestion, getErrorMessage, stripHtmlTags } from "../utils.js";
+import { formatErrorWithSuggestion, getErrorMessage } from "../utils/errors.js";
+import { stripHtmlTags } from "../utils/html.js";
 import { trackedMcpServer } from "./capabilities.js";
 
 const logger = getLogger("activitypub-mcp:tools-export");
