+++ src/mcp/tools.ts
@@ -16,7 +16,8 @@ import type { RateLimiter } from "../resilience/rate-limiter.js";
 import { validateActorIdentifier, validateDomain, validateQuery } from "../server/index.js";
 import { healthChecker } from "../telemetry/health-check.js";
 import { performanceMonitor } from "../telemetry/performance-monitor.js";
-import { formatErrorWithSuggestion, getErrorMessage, stripHtmlTags } from "../utils.js";
+import { formatErrorWithSuggestion, getErrorMessage } from "../utils/errors.js";
+import { stripHtmlTags } from "../utils/html.js";
 import { ActorIdentifierSchema, DomainSchema, QuerySchema } from "../validation/schemas.js";
 import { trackedMcpServer } from "./capabilities.js";
 import { registerExportTools } from "./tools-export.js";
